% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMaidenheadToLatLon.R
\name{convertMaidenheadToLatLon}
\alias{convertMaidenheadToLatLon}
\title{Convert Maidenhead Locator to Latitude and Longitude}
\usage{
convertMaidenheadToLatLon(maidenhead)
}
\arguments{
\item{maidenhead}{A string representing the Maidenhead locator.}
}
\value{
A named vector with `Latitude` and `Longitude` values in decimal degrees. Returns `NA` for both `Latitude` and `Longitude` if the input format is incorrect.
}
\description{
This function takes a Maidenhead locator string and converts it into
geographic coordinates (latitude and longitude). The Maidenhead locator
system is commonly used by amateur radio operators. The function supports
locators of 4 or 6 characters length.
}
\examples{
convertMaidenheadToLatLon("DM07ce")
convertMaidenheadToLatLon("CM97sh")

}
\references{
Tyson, E. T. (1989) Conversion Between Geodic and Grid Locator Systems. QST, 30-43.
}
\author{
W. Kyle Hamilton
}
