% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADIF.R
\name{readADIF}
\alias{readADIF}
\title{Read ADIF File}
\usage{
readADIF(filePath)
}
\arguments{
\item{filePath}{A string specifying the path to the ADIF file to be read.}
}
\value{
A dataframe where each row represents a record from the ADIF file. Columns correspond to tags found in the ADIF records, with column names matching tag names.
}
\description{
Reads an Amateur Data Interchange Format (ADIF) file and converts it into a dataframe. Each record in the ADIF file is transformed into a row in the dataframe.
}
\examples{

df <- readADIF(system.file("extdata", "KG6BXW.adif", package="ADIF"))
head(df)

}
\author{
W. Kyle Hamilton
}
