% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCabrillo.R
\name{readCabrillo}
\alias{readCabrillo}
\title{Read a Cabrillo File}
\usage{
readCabrillo(filePath, includeHeader = FALSE)
}
\arguments{
\item{filePath}{A string specifying the path to the Cabrillo file to be read.}

\item{includeHeader}{A logical value indicating whether to include the file's header information in the output. If TRUE, the function returns a list with two elements: `Header` containing header information, and `QSOs` containing the QSO data as a tibble. If FALSE, only the QSO data is returned as a tibble.}
}
\value{
If `includeHeader` is FALSE, returns a tibble where each row represents a QSO. If `includeHeader` is TRUE, returns a list with two elements: `Header`, a list of header information, and `QSOs`, a tibble of QSO data.
}
\description{
Reads a Cabrillo file and extracts QSO (contact) data into a tibble. Optionally, it can also extract and return header information from the Cabrillo file.
}
\examples{
\dontrun{
qsos <- readCabrillo("KG6BXW.log")
# To include header information in the output
qsosWithHeader <- readCabrillo("KG6BXW.log", includeHeader = TRUE)
}

}
\author{
W. Kyle Hamilton
}
