% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCabrilloDirectory.R
\name{readCabrilloDirectory}
\alias{readCabrilloDirectory}
\title{Read Cabrillo Files from Directory}
\usage{
readCabrilloDirectory(directoryPath, includeHeader = FALSE)
}
\arguments{
\item{directoryPath}{The path to the directory containing Cabrillo files.}

\item{includeHeader}{Logical, indicating whether to include the file header in the output. Defaults to FALSE.}
}
\value{
A tibble containing all QSOs from the Cabrillo files in the directory, with each row representing a QSO.
}
\description{
Reads all Cabrillo (.log) files from a specified directory and compiles them into a single dataframe. Optionally includes the Cabrillo file header in the output.
}
\examples{
\dontrun{
# Assuming 'Data/' directory contains Cabrillo (.log) files
completeDF <- readCabrilloDirectory("Data/")
}

}
\author{
W. Kyle Hamilton
}
